+++ example/lib/app.dart
@@ -7,6 +7,7 @@ import 'package:example/views/events_planner_rotate_example.dart';
 import 'package:example/views/events_planner_rotate_multi_columns_example.dart';
 import 'package:example/views/events_planner_rtl_example.dart';
 import 'package:example/views/events_planner_three_days_example.dart';
+import 'package:example/views/events_planner_three_days_example2.dart';
 import 'package:flutter/material.dart';
 import 'package:infinite_calendar_view/infinite_calendar_view.dart';
 import 'package:intl/date_symbol_data_local.dart';
@@ -16,6 +17,7 @@ import 'data.dart';
 import 'enumerations.dart';
 import 'views/events_list_example.dart';
 import 'views/events_planner_draggable_events_example.dart';
+import 'views/events_timeline_example.dart';
 
 var isDarkMode = false;
 var calendarMode = Mode.day3Draggable;
@@ -116,6 +118,7 @@ class CalendarViewWidget extends StatelessWidget {
       Mode.day => PlannerOneDay(key: UniqueKey()),
       Mode.day3 => PlannerTreeDays(key: UniqueKey()),
       Mode.day3Draggable => PlannerEventsDrag(key: UniqueKey(), daysShowed: 3),
+      Mode.day3Slot => PlannerTreeDaysSlot(key: UniqueKey()),
       Mode.day7 => PlannerEventsDrag(key: UniqueKey(), daysShowed: 7),
       Mode.multiColumn => PlannerMultiColumns(key: UniqueKey()),
       Mode.multiColumn2 => PlannerMultiColumns2(key: UniqueKey()),